Rebase and get rid of TS and other deps
