--- actors/autoesa-daily/.actor/actor.json
@@ -1,12 +1,0 @@
-{
-	"actorSpecification": 1,
-	"name": "autoesa-daily",
-	"title": "Empty TypeScript project",
-	"description": "Empty project in TypeScript.",
-	"version": "0.0",
-	"buildTag": "latest",
-	"meta": {
-		"templateId": "ts-empty"
-	},
-	"dockerfile": "./Dockerfile"
-}
