--- actors/autoesa-daily/.dockerignore
@@ -1,17 +1,0 @@
-# configurations
-.idea
-.vscode
-
-# crawlee and apify storage folders
-apify_storage
-crawlee_storage
-storage
-
-# installed files
-node_modules
-
-# git folder
-.git
-
-# dist folder
-dist
