--- actors/autoesa-daily/tsconfig.json
@@ -1,15 +1,0 @@
-{
-    "extends": "@apify/tsconfig",
-    "compilerOptions": {
-        "module": "NodeNext",
-        "moduleResolution": "NodeNext",
-        "target": "ES2022",
-        "outDir": "dist",
-        "noUnusedLocals": false,
-        "skipLibCheck": true,
-        "lib": ["DOM"]
-    },
-    "include": [
-        "./src/**/*"
-    ]
-}
