+++ actors/autoesa/.actor/Dockerfile
@@ -1,0 +1,8 @@
+FROM apify/actor-node:22
+
+COPY package.json ./
+
+RUN npm --quiet set progress=false \
+ && npm install --omit=dev --omit=optional
+
+COPY . ./
