+++ actors/autoesa/.actor/actor.json
@@ -1,0 +1,6 @@
+{
+  "actorSpecification": 1,
+  "name": "autoesa",
+  "version": "1.0",
+  "buildTag": "latest"
+}
