+++ actors/autoesa/.gitignore
